# api/v1alpha1/openapi.yaml
openapi: 3.0.1
info:
  title: Migration Planner API
  version: undefined
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /
paths:
...
  /api/v1/sources/{id}:
    get:
      tags:
        - source
      description: get the specified source
      operationId: getSource
      parameters:
        - name: id
          in: path
          description: ID of the source
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: NotFound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - source
      description: update a source from inventory file
      operationId: updateSource
      parameters:
        - name: id
          in: path
          description: ID of the source
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourceUpdateOnPrem'
        required: true
      responses:
        "200":
          description: An OVA image
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: NotFound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - source
      description: delete a source
      operationId: deleteSource
      parameters:
        - name: id
          in: path
          description: ID of the source
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: NotFound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          description: NotFound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Source:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        inventory:
          $ref: '#/components/schemas/Inventory'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        onPremises:
          type: boolean
        agent:
          $ref: '#/components/schemas/Agent'
      required:
        - id
        - name
        - createdAt
        - updatedAt
        - onPremises
...
    SourceUpdateOnPrem:
      type: object
      properties:
        agent_id:
          type: string
          format: uuid
        inventory:
          $ref: '#/components/schemas/Inventory'
      required:
        - inventory
        - agent_id
...
    Error:
      properties:
        message:
          type: string
          description: Error message
      required:
        - message
...
    Inventory:
      type: object
      required:
      - vcenter
      - vms
      - infra
      properties:
        vcenter:
          type: object
          $ref: '#/components/schemas/VCenter'
        vms:
          type: object
          $ref: '#/components/schemas/VMs'
        infra:
          type: object
          $ref: '#/components/schemas/Infra'
    VCenter:
      type: object
      required:
        - id
      properties:
        id:
          type: string
    VMs:
      type: object
      required:
      - total
      - totalMigratable
      - cpuCores
      - ramGB
      - diskGB
      - diskCount
      - os
      - powerStates
      - notMigratableReasons
      - migrationWarnings
      properties:
        total:
          type: integer
        totalMigratable:
          type: integer
        totalMigratableWithWarnings:
          type: integer
        cpuCores:
          type: object
          $ref: '#/components/schemas/VMResourceBreakdown'
        ramGB:
          type: object
          $ref: '#/components/schemas/VMResourceBreakdown'
        diskGB:
          type: object
          $ref: '#/components/schemas/VMResourceBreakdown'
        diskCount:
          type: object
          $ref: '#/components/schemas/VMResourceBreakdown'
        powerStates:
          type: object
          additionalProperties:
            type: integer
        os:
          type: object
          additionalProperties:
            type: integer
        notMigratableReasons:
          type: object
          $ref: '#/components/schemas/MigrationIssues'
        migrationWarnings:
          type: object
          $ref: '#/components/schemas/MigrationIssues'
    Infra:
      type: object
      required:
      - totalHosts
      - totalClusters
      - hostsPerCluster
      - hostPowerStates
      - networks
      - datastores
      properties:
        totalHosts:
          type: integer
        totalClusters:
          type: integer
        hostsPerCluster:
          type: array
          items:
            type: integer
        hostPowerStates:
          type: object
          additionalProperties:
            type: integer
        networks:
          type: array
          items:
            type: object
            required:
              - type
              - name
            properties:
              type:
                type: string
                enum: [standard, distributed, dvswitch, unsupported]
              name:
                type: string
              vlanId:
                type: string
              dvswitch:
                type: string
        datastores:
          type: array
          items:
            type: object
            required:
              - type
              - totalCapacityGB
              - freeCapacityGB
            properties:
              type:
                type: string
              totalCapacityGB:
                type: integer
              freeCapacityGB:
                type: integer
    VMResourceBreakdown:
      type: object
      required:
      - histogram
      - total
      - totalForMigratable
      - totalForMigratableWithWarnings
      - totalForNotMigratable
      properties:
        total:
          type: integer
        totalForMigratable:
          type: integer
        totalForMigratableWithWarnings:
          type: integer
        totalForNotMigratable:
          type: integer
        histogram:
          type: object
          required:
          - minValue
          - step
          - data
          properties:
            minValue:
              type: integer
            step:
              type: integer
            data:
              type: array
              items:
                type: integer
    MigrationIssues:
      type: array
      items:
        type: object
        required:
        - label
        - assessment
        - count
        properties:
          label:
            type: string
          assessment:
            type: string
          count:
            type: integer
    Agent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [not-connected, waiting-for-credentials, error, gathering-initial-inventory, up-to-date, source-gone]
        statusInfo:
          type: string
        credentialUrl:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        version:
          type: string
      required:
        - id
        - status
        - statusInfo
        - credentialUrl
        - createdAt
        - updatedAt
        - version
...